Integrate feedback
